% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_format}
\alias{data_format}
\alias{data_format.network}
\title{Generate "Format" section of data documentation}
\usage{
data_format(object, ...)

\method{data_format}{network}(object, ...)
}
\arguments{
\item{object}{data object to be described see "Methods" section below for
available methods}
}
\value{
Character scalar with the Markdown text.
}
\description{
Generates text in Markdown for the "Format" section (without the
section itself). It is a generic function with methods defined below.
}
\section{Methods (by class)}{
\itemize{
\item \code{network}: Create a paragraph providing information on the size
of the node set, the edge set and the node and edge attributes.
}}

\concept{data documentation generators}
