% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoli.R
\docType{data}
\name{ecoli}
\alias{ecoli}
\alias{ecoli1}
\alias{ecoli2}
\title{Two versions of an E. Coli network dataset}
\format{
\code{ecoli1} - Directed \code{\link[network:network]{network}} object with 423 nodes and 519 edges.
Vertex attributes: \code{self}.

\code{ecoli2} - Undirected \code{\link[network:network]{network}} object with 418 nodes and 519 edges.
Vertex attributes: \code{self}.
}
\source{
The data set is based on the RegulonDB network (Salgado et al, 2001)
and was modified by Shen-Orr et al (2002).
}
\description{
This network data set comprises two versions of a biological network in
which the nodes are operons in \emph{Escherichia Coli} and a directed edge
from one node to another indicates that the first encodes the transcription
factor that regulates the second.
}
\details{
The network object \code{ecoli1} is directed, with 423 nodes and 519
arcs. The object \code{ecoli2} is an undirected version of the same network, in
which all arcs are treated as edges and the five isolated nodes (which
exhibit only self-regulation in \code{ecoli1}) are removed, leaving 418 nodes.
}
\section{Licenses and Citation}{
 When publishing results obtained using this
data set, the original authors (see sections Source and References) should
be cited, along with this \code{R} package.

If the section Source of the data set does not specified otherwise, this data
set is protected by the Creative Commons License
\url{https://creativecommons.org/licenses/by-nc-nd/2.5/}.
}

\references{
Salgado et al (2001), Regulondb (version 3.2): Transcriptional Regulation
and Operon Organization in Escherichia Coli K-12, \emph{Nucleic Acids
Research}, 29(1): 72-74.

Shen-Orr et al (2002), Network Motifs in the Transcriptional Regulation
Network of Escerichia Coli, \emph{Nature Genetics}, 31(1): 64-68.

Saul, Z. M., & Filkov, V. (2007). Exploring biological network structure
using exponential random graph models. Bioinformatics, 23(19), 2604-2611.
}
\seealso{
\code{\link[network:network]{network::network()}}, \code{\link[network:plot.network]{network::plot.network()}}

Other directed networks: 
\code{\link{emon}},
\code{\link{faux.desert.high}},
\code{\link{faux.dixon.high}},
\code{\link{g4}},
\code{\link{newcomb}},
\code{\link{samplk}},
\code{\link{sampson}},
\code{\link{tribes}}

Other undirected networks: 
\code{\link{davis}},
\code{\link{faux.magnolia.high}},
\code{\link{faux.mesa.high}},
\code{\link{florentine}},
\code{\link{kapferer}},
\code{\link{molecule}},
\code{\link{zach}}
}
\concept{directed networks}
\concept{undirected networks}
\keyword{datasets}
