% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g4.R
\docType{data}
\name{g4}
\alias{g4}
\title{Goodreau's four node network as a "network" object}
\format{
It is a directed network of four nodes and five ties stored as a
\code{\link[network:network]{network}} object.
}
\source{
Steve Goodreau
}
\usage{
g4
}
\description{
This is an example thought of by Steve Goodreau.
}
\details{
It is interesting because the maximum likelihood estimator of the model with
out degree 3 in it exists, but the maximum psuedolikelihood estimator does
not.
}
\section{Licenses and Citation}{
 If the section Source of this page does not
specify otherwise, this data set is protected by the Creative Commons
License \url{https://creativecommons.org/licenses/by-nc-nd/2.5/}.

When publishing results obtained using this data set, the original authors
(see sections Source and/or References) should be cited, along with this
\code{R} package. To cite this package please use the following:

Handcock M, Hunter D, Butts C, Goodreau S, Krivitsky P, Morris M,
Bojanowski M (2021). \emph{statnet.data: Network Datasets for the Statnet
Suite}. R package version 0.1-0, <URL: https://statnet.org>.
}

\examples{
\dontrun{
library(ergm)
summary(ergm(g4 ~ odegree(3), estimate="MPLE"))
summary(ergm(g4 ~ odegree(3), control=control.ergm(init=0)))
}

}
\seealso{
Other directed networks: 
\code{\link{ecoli}},
\code{\link{emon}},
\code{\link{faux.desert.high}},
\code{\link{faux.dixon.high}},
\code{\link{newcomb}},
\code{\link{samplk}},
\code{\link{sampson}},
\code{\link{tribes}}
}
\concept{directed networks}
\keyword{datasets}
