% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecule.R
\docType{data}
\name{molecule}
\alias{molecule}
\title{Synthetic network with 20 nodes and 28 edges}
\format{
Undirected \code{\link[network:network]{network}} object with 20 nodes and 28
edges
}
\usage{
molecule
}
\description{
This is a synthetic network of 20 nodes that is used as an example within
the \code{ergm} documentation. It has an interesting elongated shape
\itemize{
\item reminencent of a chemical molecule.
}
}
\section{Licenses and Citation}{
 If the section Source of this page does not
specify otherwise, this data set is protected by the Creative Commons
License \url{https://creativecommons.org/licenses/by-nc-nd/2.5/}.

When publishing results obtained using this data set, the original authors
(see sections Source and/or References) should be cited, along with this
\code{R} package. To cite this package please use the following:

Handcock M, Hunter D, Butts C, Goodreau S, Krivitsky P, Morris M,
Bojanowski M (2021). \emph{statnet.data: Network Datasets for the Statnet
Suite}. R package version 0.1-0, <URL: https://statnet.org>.
}

\seealso{
Other undirected networks: 
\code{\link{davis}},
\code{\link{ecoli}},
\code{\link{faux.magnolia.high}},
\code{\link{faux.mesa.high}},
\code{\link{florentine}},
\code{\link{kapferer}},
\code{\link{zach}}
}
\concept{undirected networks}
\keyword{datasets}
