% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribes.R
\docType{data}
\name{tribes}
\alias{tribes}
\title{Read Highland Tribes}
\format{
An undirected \code{\link[network:network]{network}} object with no loops, having the following
vertex attributes:
\itemize{
\item \code{vertex.names} -- Character attribute with names of tribes.
}

Edge attributes:
\itemize{
\item \code{pos} -- Logical attribute indicating an alliance relationship.
\item \code{neg} -- Logical attribute indicating a hostile relationship ("rova")
\item \code{sign} -- Numeric attribute coding -1 for enmity, 0 for no relationship, and 1 for alliance.
\item \code{sign.012} -- Numeric attribute coding 0 for enmity, 1 for no relationship, and 2 for alliance.
}

Because of limitations of \code{network} objects, the object itself is a complete
graph, and is thus meaningless if used directly or plotted.
}
\source{
\url{http://vlado.fmf.uni-lj.si/pub/networks/data/UciNet/UciData.htm#gama},
with corrections from Read (1954).
}
\description{
A network of political alliances and enmities among the 16 Gahuku-Gama
sub-tribes of Eastern Central Highlands of New Guinea, documented by Read
(1954). This network shows 3 clusters.
}
\section{Licenses and Citation}{
 When publishing results obtained using this
data set, the original authors (see sections Source and References) should
be cited, along with this \code{R} package.

If the section Source of the data set does not specified otherwise, this data
set is protected by the Creative Commons License
\url{https://creativecommons.org/licenses/by-nc-nd/2.5/}.
}

\examples{
\dontrun{
library(latentnet)
# Only model positive ties:
tribes.fit < -ergmm(tribes ~ euclidean(d=2, G=3), response="pos")
# Edge color must be set manually, for green ties to represent alliance
# and for red ties to represent enmity.
plot(tribes.fit,
  edge.col = as.matrix(tribes, "pos", m="a") * 3 + as.matrix(tribes,"neg",m="a")*2,
  pie = TRUE)
# Model both positive and negative ties:
tribes.fit3 < -ergmm(tribes ~ euclidean(d=2, G=3), response="sign.012",
                   family="binomial.logit", fam.par=list(trials=2))
# Edge color must be set manually, for green ties to represent alliance
# and for red ties to represent enmity.
plot(tribes.fit3,
  edge.col = as.matrix(tribes,"pos",m="a")*3+as.matrix(tribes,"neg",m="a")*2,
  pie=TRUE)
}
}
\references{
Taken from UCINET IV, which cites the following:

Hage P. and Harary F. (1983). Structural models in anthropology. Cambridge:
Cambridge University Press. (See p 56-60).  Read K. (1954). Cultures of the
central highlands, New Guinea. Southwestern Journal of Anthropology, 10,
1-43.
}
\seealso{
Other directed networks: 
\code{\link{ecoli}},
\code{\link{emon}},
\code{\link{faux.desert.high}},
\code{\link{faux.dixon.high}},
\code{\link{g4}},
\code{\link{newcomb}},
\code{\link{samplk}},
\code{\link{sampson}}
}
\concept{directed networks}
\keyword{cluster}
\keyword{graphs}
\keyword{multivariate}
