% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zach.R
\docType{data}
\name{zach}
\alias{zach}
\title{Karate club social network of Zachary (1977)}
\format{
The data are represented as a \code{\link[network:network]{network}} object, with an
edge attribute
\itemize{
\item \code{contexts} -- the number of contexts of interaction for that pair of
actors.
}

In addition, the following vertex attributes are
provided:
\itemize{
\item \code{club} -- the club in which the actor ended up
\item \code{faction} -- faction alignment of the actor as recorded by Zachary
\item \code{faction.id} -- faction alignment coded numerically, as -2 (strongly Mr.
Hi's), -1 (weakly Mr. Hi's), 0 (neutral), +1 (weakly John's), and +2
(strongly John's)
\item \code{role} -- role of the actor in the network (Instructor, Member, or
President)
}
}
\source{
Zachary, WW (1977). An Information Flow Model for Conflict and
Fission in Small Groups. \emph{Journal of Anthropological Research}, 33(4),
452-473.

Sociomatrix in machine-readable format was retrieved from
\url{http://vlado.fmf.uni-lj.si/pub/networks/data/ucinet/ucidata.htm}.
}
\usage{
zach
}
\description{
Zachary (1977) reported observations of social relations in a university
karate club
}
\details{
Zachary (1977) reported observations of social relations in a university
karate club, with membership that varied between 50 and 100, of whom 34
individuals: 32 ordinary club members and officers, the club president ("John
A."), and the part-time instructor ("Mr. Hi"); consistently interacted
outside of the club. Over the course of the study, the club divided into two
factions, and, ultimately, split into two clubs, one led by Hi and the other
by John and the original club's officers. The split was driven by a
disagreement over whether Hi could unilaterally change the level of
compensation for his services.

Zachary identifies the faction with which each of the 34 actors was aligned
and how strongly and reports, for each pair of actors, the count of social
contexts in which they interacted. The 8 contexts recorded were
\itemize{
\item academic classes at the university
\item Hi's private karate studio in his night classes
\item Hi's private karate studio where he taught on weekends
\item student-teaching at Hi's studio
\item the university rathskeller (bar) located near the karate club
\item a bar located near the university campus
\item open karate tournaments in the area; and
\item intercollegiate karate tournaments.
}

The highest number of contexts of interaction for a pair of individuals that
was observed was 7.
}
\section{Licenses and Citation}{
 If the section Source of this page does not
specify otherwise, this data set is protected by the Creative Commons
License \url{https://creativecommons.org/licenses/by-nc-nd/2.5/}.

When publishing results obtained using this data set, the original authors
(see sections Source and/or References) should be cited, along with this
\code{R} package. To cite this package please use the following:

Handcock M, Hunter D, Butts C, Goodreau S, Krivitsky P, Morris M,
Bojanowski M (2021). \emph{statnet.data: Network Datasets for the Statnet
Suite}. R package version 0.1-0, <URL: https://statnet.org>.
}

\examples{
if(requireNamespace("network", quietly=TRUE)){
oldpal <- palette()
palette(gray((1:8)/8))
plot(zach, vertex.col="role", displaylabels=TRUE, edge.col="contexts")
palette(oldpal)
}

\dontrun{
# Fit a binomial-reference ERGM.
zach.fit1 <- ergm(zach~nonzero+sum+nodefactor("role",base=2)+absdiffcat("faction.id"),
                  response="contexts", reference=~Binomial(8),
                  control=control.ergm(MCMLE.trustregion=1000))
mcmc.diagnostics(zach.fit1)
summary(zach.fit1)

# This is much slower.
zach.fit2 <- ergm(zach~nonzero+sum+nodefactor("role",base=2)+transitiveties,
                  response="contexts", reference=~Binomial(8),
                  control=control.ergm(MCMLE.trustregion=1000),
                  eval.loglik=FALSE)
mcmc.diagnostics(zach.fit2)
summary(zach.fit2)
}

}
\references{
Zachary, WW (1977). An Information Flow Model for Conflict and
Fission in Small Groups. \emph{Journal of Anthropological Research}, 33(4),
452-473.
}
\seealso{
Other undirected networks: 
\code{\link{davis}},
\code{\link{ecoli}},
\code{\link{faux.magnolia.high}},
\code{\link{faux.mesa.high}},
\code{\link{florentine}},
\code{\link{kapferer}},
\code{\link{molecule}}
}
\concept{undirected networks}
\keyword{datasets}
