% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g4.R
\docType{data}
\name{g4}
\alias{g4}
\title{Goodreau's four node network as a "network" object}
\format{
It is a directed network of four nodes and five ties stored as a
\code{\link[network:network]{network}} object.
}
\source{
Steve Goodreau
}
\usage{
g4
}
\description{
This is an example thought of by Steve Goodreau.
}
\details{
It is interesting because the maximum likelihood estimator of the model with
out degree 3 in it exists, but the maximum psuedolikelihood estimator does
not.
}
\section{Licenses and Citation}{
 When publishing results obtained using this
data set, the original authors (see sections Source and References) should
be cited, along with this \code{R} package.
}

\examples{

summary(ergm(g4 ~ odegree(3), estimate="MPLE"))
summary(ergm(g4 ~ odegree(3), control=control.ergm(init=0)))

}
\seealso{
\code{\link[network:network]{network::network()}}, \code{\link[network:plot.network]{network::plot.network()}}

Other directed networks: 
\code{\link{ecoli}},
\code{\link{emon}},
\code{\link{faux.desert.high}},
\code{\link{faux.dixon.high}},
\code{\link{samplk}},
\code{\link{sampson}},
\code{\link{tribes}}
}
\concept{directed networks}
\keyword{datasets}
