% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecule.R
\docType{data}
\name{molecule}
\alias{molecule}
\title{Synthetic network with 20 nodes and 28 edges}
\format{
Undirected \code{\link[network:network]{network}} object with 20 nodes and 28
edges
}
\usage{
molecule
}
\description{
This is a synthetic network of 20 nodes that is used as an example within
the \code{ergm} documentation. It has an interesting elongated shape
\itemize{
\item reminencent of a chemical molecule.
}
}
\section{Licenses and Citation}{
 When publishing results obtained using this
data set, the original authors (see sections Source and References) should
be cited, along with this \code{R} package.
}

\seealso{
\code{\link[network:network]{network::network()}}, \code{\link[network:plot.network]{network::plot.network()}}

Other undirected networks: 
\code{\link{ecoli}},
\code{\link{faux.magnolia.high}},
\code{\link{faux.mesa.high}},
\code{\link{florentine}},
\code{\link{kapferer}}
}
\concept{undirected networks}
\keyword{datasets}
